Web funcionando

+++ build/CMakeFiles/clean_additional.cmake
@@ -1,15 +1,17 @@
 # Additional clean files
+cmake_minimum_required(VERSION 3.16)
 
-file(REMOVE_RECURSE
-  "bootloader/bootloader.bin"
-  "bootloader/bootloader.elf"
-  "bootloader/bootloader.map"
-  "config/sdkconfig.cmake"
-  "config/sdkconfig.h"
+if("${CONFIG}" STREQUAL "" OR "${CONFIG}" STREQUAL "")
+  file(REMOVE_RECURSE
+  "bootloader\\bootloader.bin"
+  "bootloader\\bootloader.elf"
+  "bootloader\\bootloader.map"
+  "config\\sdkconfig.cmake"
+  "config\\sdkconfig.h"
   "controle-quarto.bin"
   "controle-quarto.map"
-  "esp-idf/esptool_py/flasher_args.json.in"
-  "esp-idf/mbedtls/x509_crt_bundle"
+  "esp-idf\\esptool_py\\flasher_args.json.in"
+  "esp-idf\\mbedtls\\x509_crt_bundle"
   "flash_app_args"
   "flash_bootloader_args"
   "flash_project_args"
@@ -18,4 +20,5 @@ file(REMOVE_RECURSE
   "ldgen_libraries.in"
   "project_elf_src_esp32.c"
   "x509_crt_bundle.S"
-)
+  )
+endif()
